Fix randomness

+++ main_fed.py
@@ -17,7 +17,11 @@ import os
 import pdb
 import time
 
+
 if __name__ == '__main__':
+    # reproduce randomness
+    torch.manual_seed(1001)
+    np.random.seed(1001)
     # parse args
     args = args_parser()
     args.device = torch.device('cuda:{}'.format(args.gpu) if torch.cuda.is_available() and args.gpu != -1 else 'cpu')
@@ -28,16 +32,30 @@ if __name__ == '__main__':
         os.makedirs(os.path.join(base_dir, 'fed'), exist_ok=True)
 
     dataset_train, dataset_test, dict_users_train, dict_users_test = get_data(args)
+    '''
     print('type: ', type(dataset_test))
     print('len: ', len(dataset_test))
-
-    dict_save_path = os.path.join(base_dir, 'dict_users.pkl')
-    with open(dict_save_path, 'wb') as handle:
-        pickle.dump((dict_users_train, dict_users_test), handle)
-
+    local_data_size = []
+    for idx in range(args.num_users):
+        local_data_size.append(len(dict_users_train[idx]))
+    print('local dataset size: ', local_data_size.sort())
+    '''
+    shard_path = './save/{}/{}_iid{}_num{}_C{}_le{}/shard{}/'.format(
+        args.dataset, args.model, args.iid, args.num_users, args.frac, args.local_ep, args.shard_per_user)
+    dict_save_path = os.path.join(shard_path, 'shared_dict_users.pkl')
+    if os.path.exists(dict_save_path): # use old one
+        print('Local data already exist!')
+        with open(dict_save_path, 'rb') as handle:
+            (dict_users_train, dict_users_test) = pickle.load(handle)
+    else:
+        print('Re dispatch data to local!')
+        with open(dict_save_path, 'wb') as handle:
+            pickle.dump((dict_users_train, dict_users_test), handle)
+        
     # build model
     net_glob = get_model(args)
     net_glob.train()
+    #print(list(net_glob.layer_hidden1.weight)[0])
 
     # training
     results_save_path = os.path.join(base_dir, 'fed/results.csv')
@@ -60,16 +78,27 @@ if __name__ == '__main__':
         loss_locals = []
         m = max(int(args.frac * args.num_users), 1) # num of selected clients
         idxs_users = np.random.choice(range(args.num_users), m, replace=False)
-        print("Round {}, lr: {:.6f}, {}".format(iter, lr, idxs_users))
 
+
+        print("Round {}, lr: {:.6f}, {}".format(iter, lr, idxs_users))
+        '''
+        for idx in idxs_users:
+            print('user: ', idx, '==================================')
+            print('dict_users_train[idx]: ', type(dict_users_train[idx]), len(dict_users_train[idx]))
+            #print(dict_users_train[idx])
+        '''
+        
         for idx in idxs_users: # iter over selected clients
             t_leps_bgin = time.time()
 
             local = LocalUpdate(args=args, dataset=dataset_train, idxs=dict_users_train[idx])
             net_local = copy.deepcopy(net_glob)
 
-            w_local, loss = local.train(net=net_local.to(args.device))
+            w_local, loss = local.train(net=net_local.to(args.device), lr=lr)
             loss_locals.append(copy.deepcopy(loss))
+            # loss: a float, avg loss over local epochs over batches
+            #print('loss: ', loss)
+
 
             if w_glob is None:
                 w_glob = copy.deepcopy(w_local)
@@ -80,11 +109,11 @@ if __name__ == '__main__':
             t_leps_end = time.time()
             time_locals.append(t_leps_end - t_leps_bgin)
 
-        lr *= args.lr_decay
+        lr *= args.lr_decay # default: no decay
 
         # update global weights
         for k in w_glob.keys():
-            w_glob[k] = torch.div(w_glob[k], m)
+            w_glob[k] = torch.div(w_glob[k], m) # calculate avg by dividing m
 
         # copy weight to net_glob
         net_glob.load_state_dict(w_glob)
@@ -118,11 +147,11 @@ if __name__ == '__main__':
             final_results = np.array(results)
             final_results = pd.DataFrame(final_results, columns=['epoch', 'loss_avg', 'loss_test', 'acc_test', 'best_acc', 'time_local_avg', 'time_glob'])
             final_results.to_csv(results_save_path, index=False)
-
+        '''
         if (iter + 1) % 50 == 0:
             best_save_path = os.path.join(base_dir, 'fed/best_{}.pt'.format(iter + 1))
             model_save_path = os.path.join(base_dir, 'fed/model_{}.pt'.format(iter + 1))
             torch.save(net_best.state_dict(), best_save_path)
             torch.save(net_glob.state_dict(), model_save_path)
-
+        '''
     print('Best model, iter: {}, acc: {}'.format(best_epoch, best_acc))
